refactor: update import paths for Entry type in blog components
