+++ apps/blog/src/components/article-header.astro
@@ -4,7 +4,7 @@ import { format } from "date-fns";
 import { LuCalendarDays, LuTags, LuDot, LuCoffee } from "react-icons/lu";
 import { cn, slug } from "@mogeko/utils";
 import type { HTMLAttributes } from "astro/types";
-import type { Entry } from "@/data";
+import type { Entry } from "@/utils";
 
 const { class: className, data, ...props } = Astro.props;
