+++ apps/blog/src/components/post-group.astro
@@ -4,7 +4,7 @@ import { format } from "date-fns";
 import { MdStar } from "react-icons/md";
 import { cn } from "@mogeko/utils";
 import type { HTMLAttributes } from "astro/types";
-import type { Entry } from "@/data";
+import type { Entry } from "@/utils";
 
 const { class: className, heading, posts: articles } = Astro.props;
