+++ apps/blog/src/pages/posts/[...slug]/index.astro
@@ -5,7 +5,7 @@ import Horizontal from "@/components/mdx/hr.astro";
 import Image from "@/components/mdx/img.astro";
 import Preformatted from "@/components/mdx/pre.astro";
 import ArrowLink from "@/components/arrow-link.astro";
-import { getEntries } from "@/data";
+import { getEntries } from "@/utils";
 import type { GetStaticPaths } from "astro";
 
 export const getStaticPaths = (async () => {
