+++ apps/blog/src/pages/tags/[...tag]/index.astro
@@ -3,9 +3,9 @@ import Layout from "@/layouts/common-layout.astro";
 import PostIndex from "@/components/post-index.astro";
 import Icon from "@/components/ui/icon.astro";
 import { slug } from "@mogeko/utils";
-import { getEntries } from "@/data";
+import { getEntries } from "@/utils";
 import type { GetStaticPaths } from "astro";
-import type { Entry } from "@/data";
+import type { Entry } from "@/utils";
 
 export const getStaticPaths = (async () => {
   const tagsWithPosts = (await getEntries()).reduce<Record<string, Entry[]>>(
